requirements.txt
fastapi
uvicorn[standard]
python-dotenv
cloudinary
python-multipart